feat: coverage increased to 80.5%
